requirements.txt
django==4.0.10
djangorestframework==3.13
mysqlclient==2.2.0
django-cors-headers==4.3.0